+++ lib/run/coordinator.js
@@ -1,9 +1,9 @@
 import { when } from '@bigtest/convergence';
 
 import logger from '../util/logger';
-import ClientServer from './client';
-import SocketServer from './sockets';
-import ProxyServer from './proxy';
+import ClientServer from './servers/client';
+import SocketServer from './servers/sockets';
+import ProxyServer from './servers/proxy';
 
 import ReporterManager from './reporters';
 import LauncherManager from './browsers';
