+++ lib/run/servers/client.js
@@ -1,5 +1,5 @@
 import path from 'path';
-import WebServer from './server';
+import WebServer from './web';
 
 /**
  * Client UI server that simply serves the client directory and any
@@ -12,7 +12,7 @@ export default class ClientServer extends WebServer {
     super(options);
 
     this.serve('/', require('express').static(
-      path.join(__dirname, 'client')
+      path.join(__dirname, '../client')
     ));
   }
 }
