+++ lib/run/servers/proxy.js
@@ -1,4 +1,4 @@
-import WebServer from './server';
+import WebServer from './web';
 
 const {
   assign,
